% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san_stochastic.R
\name{find_time_step}
\alias{find_time_step}
\title{Determine a sensible time step}
\usage{
find_time_step(rate, p_cutoff)
}
\arguments{
\item{rate}{the maximal rate in the model}

\item{p_cutoff}{the maximal allowed probability of two sequential events
within one time step}
}
\description{
The time step is chosen to reduce the probability of two sequential
events (i.e. where the second event affects a cell created or destroyed
by the first event) below \code{p_cutoff}
}
