% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SAN.RATENAMES}
\alias{SAN.RATENAMES}
\title{The names of the rate parameters of the deterministic and
stochastic SAN model}
\format{
An object of class \code{character} of length 6.
}
\usage{
SAN.RATENAMES
}
\description{
The names of the rate parameters of the deterministic and
stochastic SAN model
}
\keyword{datasets}
