% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san_deterministic.R
\name{san_deterministic}
\alias{san_deterministic}
\title{Computes the number of S-, A- and N-cells under the deterministic SAN model}
\usage{
san_deterministic(s0, rates, samples_per_day = 1)
}
\arguments{
\item{s0}{the initial number of S-cells}

\item{rates}{a `data.table` with columns `Tmax`, `r_S`, `r_0`, `r_R`, `r_A`, `r_N`
and `r_D`  and monotonically increasing values in the column `Tmax`.}

\item{samples_per_day}{the number of (equally spaced) times point per day at
which to output the cell counts}
}
\value{
a `data.table` with columns `t`, `S`, `A`, `N` containing the cell counts
        at each day from \eqn{t=0} to \eqn{t=T}, where \eqn{T} is the value of
        `Tmax` in the last row of the rates table. For each day, the table contains
        `samples_per_day` rows with equally spaced evaluation times within that day.
}
\description{
The deterministic SAN model describes the *average* behavior of the stochastic
SAN model, and is the defined by the ODE
\deqn{S' = (r_S - r_0 - r_R - r_A)\cdot S}
\deqn{A' = r_A\cdot S - r_D\cdot A}
\deqn{N' = r_R\cdot S + r_N\cdot A}
This function computes \eqn{(S(t), A(t), N(t))} from \eqn{t=0} onwards
given \eqn{s_0=S(0)} and time-dependent but piece-wise constant rates
\eqn{r_S}, \eqn{r_0}, \eqn{r_R}, \eqn{r_A}, \eqn{r_SN}, \eqn{r_D}.
}
\details{
The picewise constant rates are specified through the table `rates` with
columns `Tmax`, `r_S`, `r_0`, `r_R`, `r_A`, `r_N`. The parameters in each row
are in effect between the previous row's `Tmax` and the row's own `Tmax`.
}
