% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san_deterministic.R
\name{san_deterministic_expression}
\alias{san_deterministic_expression}
\title{Returns the analytic expression for the specified parameter regime of the deterministic SAN model}
\usage{
san_deterministic_expression(regime)
}
\arguments{
\item{regime}{the parameter regime as returned by `san_deterministic_regime`}
}
\value{
an expression in the variables `s0`, `a0`, `n0`, `r_S`, `r_0`, `r_R`, `r_A`, `r_N`, `r_D` and `times`.
}
\description{
If the goal is simply to evaluate the deterministic SAN model, `san_deterministic` should
be used which allows rates to vary over time, provides a user-friedly API,
and checks parameters for validity.
}
