% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san_stochastic.R
\name{san_stochastic}
\alias{san_stochastic}
\title{Simulate the stochastic SAN model}
\usage{
san_stochastic(L, s0 = 1, rates, samples_per_day = 1, p_cutoff = 0.001)
}
\arguments{
\item{s0}{the initial number of S-cells}

\item{rates}{a `data.table` with columns `Tmax`, `r_S`, `r_0`, `r_R`, `r_A`, `r_N`
and `r_D`  and monotonically increasing values in the column `Tmax`.}

\item{samples_per_day}{the number of (equally spaced) times point per day at
which to output the cell counts}

\item{p_cutoff}{the maximal allowed probability of two sequential events
within one time step}
}
\value{
a `data.table` with columns `t`, `S`, `A`, `N` containing the cell counts
        at each day from \eqn{t=0} to \eqn{t=T}, where \eqn{T} is the value of
        `Tmax` in the last row of the rates table. For each day, the table contains
        `samples_per_day` rows with equally spaced evaluation times within that day. 

While the model as defined above is time-continuous, this function simulates
the model in discrete time steps. The parameter `p_cutoff` controls the accuracy
of this discretization, see \code{\link[=func]{find_time_step()}}
}
\description{
The stochastic SAN model describes the stochastic behavior of lineages
consisting of S, A and N-cells which can undergo the following conversions
in a memoryless fashion, making the model a Markov process.
}
\details{
| event                 | rate      |
| \eqn{S \to S S}       | \eqn{r_S} |
| \eqn{S \to \emptyset} | \eqn{r_0} |
| \eqn{S \to N}         | \eqn{r_R} |
| \eqn{S \to A}         | \eqn{r_A} |
| \eqn{A \to A N}       | \eqn{r_N} |
| \eqn{A \to N}         | \eqn{r_D} |

The picewise constant rates are specified through the table `rates` with
columns `Tmax`, `r_S`, `r_0`, `r_R`, `r_A`, `r_N`. The parameters in each row
are in effect between the previous row's `Tmax` and the row's own `Tmax`.
}
